chore: extract shared dh-files vite plugin
vite.config.ts and vitest.config.ts each carried an identical copy of
dhFilesPlugin. It now lives in dh-files-plugin.ts, imported by both, so
the .dh/.dhkit loader cannot drift between app builds and tests. Added
to the tsconfig include so it stays type-checked.
